refactor: extract composable
